make GameBoard's receiveAttack() and this.#grid
function will take a coordinate and respond with either a hit or a miss
of a ship. This will be crucial when playing battleship so user knowns
the outcome of there choice. this commits is including private property
grid which holds all the positions and has the ships inside
this version passes the current tests