prevent placing out of bounds & reset at drag end
now the board stops the drag when the ships are being placed
outside the board, it does not allow it to happen. The reset is so
there is no error when the drag ends outside the board. There was an
error after the drag ended outside the board, when they tried to
move the ship it would shift a square away from where it was placed